CopyFromImg_source_code/requirements.txt
clipboard==0.0.4
keyboard==0.13.5
Pillow==9.3.0
pyinstaller==5.7.0
pynput==1.7.6
pytesseract==0.3.10

